% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDataExtractor.R
\name{metaDataExtractor}
\alias{metaDataExtractor}
\title{Extract meta-data for image files}
\usage{
metaDataExtractor(whatSeperates, matchedXml)
}
\arguments{
\item{whatSeperates}{n element vector of elements in the imagenames given (eg. 0001_rightwing_kd_rrg c("num","wing","imagerInitials","genotype"))}

\item{matchedXml}{list object containing xml and xlif files produced by xmlTiffMatcher when run without errors}
}
\description{
Extracts metadata from xlif and xml files that have matching images.
}
\examples{
whatSeperates<-c("num","wing","imagerInitials","knf")
}
\author{
Kevin Doheny
}
