% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataLandmarkCombine.R
\name{metadataLandmarkCombine}
\alias{metadataLandmarkCombine}
\title{Combines metadata and landmark files produced by tpsDig}
\usage{
metadataLandmarkCombine(metaData, tpsFileName, dir, outputFileName)
}
\arguments{
\item{metaData}{list of extracted metadata files}

\item{tpsFileName}{character string (1) name of tps file (no suffix)}

\item{dir}{character string (1) path to directory where tps file exists}

\item{outputFileName}{character string (1) output file name (no suffix)}
}
\value{
.asc and .csv datafiles with landmarks coded for the Wings4 program
}
\description{
Combines metadata and landmark files produced by tpsDig and outputs into .csv and .asc formats.
}
\examples{
tpsFileName<-"asc5.TPS"
outputFileName<-"ForWings"
dir<-"A:/Github/microscopyGithub/test2"
metadataLandmarkCombine(metaData, tpsFile,dir,outputFileName)
}
