% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlTiffMatcher.R
\name{xmlTiffMatcher}
\alias{xmlTiffMatcher}
\title{Matches .xml files with .tiff files in processing pipeline}
\usage{
xmlTiffMatcher(meta, tiffLocate)
}
\arguments{
\item{meta}{character string(1) path to directory of metadatafiles}

\item{tiffLocate}{character string(1) path to directory of image files}
}
\value{
A list of xml files with matching tifs where xml entries occur before xlif entries.
}
\description{
xml and .tiff matcher.
Makes sure that every image taken has an associated xml file with scale data.
Outputs a list of .xml and xlif files for information extraction.
Outputs a vector of .xlif files for information extraction.
}
\examples{
meta<-"A:/PhD/HalfSib/D.suzukiiHS/allData/leicametadata"
tiflocate<-"C:/Users/sztfl/sztfl/Imaging/suzukii_binom_mc_feb2022_block1/"
}
